+++ internal/producer/src/producer/genesis_hashes.txt
@@ -1,0 +1,50 @@
+bd224623c3314cf8202e688c61c2bca83f47a67e34b0d263eabecfd4651f9cb4
+666a165cd279a05fe962c1e57e12aa11ef5ce463349bd144193cc59293d24998
+b8b8881e8e65015c721bdc7997fdc44b19796a9cbb4b9d463d726ece708a981b
+4910a1eb90b0b00488136cea7ad4ee839d86ece2fb0ed832ee80c9a2f050ad8f
+bc8de20589b4af04b41d54b55f00ffd297ad62fa21a7c692a425831221a9b9e5
+73bcb757f4c73998b721e0540d3c0047d5108846c69fabe617ec42c75d59fb02
+cb1637b78877dc31ffa21cf39906dd003eb345fcb2153a27e95aac93569f8124
+ee16dfaf5e189fdb2833dc52c2b2e0d7517c6f8bf1842292bab9a45eab50bb04
+0caaea46d6483381a91c468f30813afdaf5f8e8683aa9388b2b1984e91880714
+da7efacea1aef2a59802e96436ed24028439125e130e08ba214671ad28a47706
+fc2692667d9375007a05affbcc4143be52638e8b45f2c7ca4b3476af4d71a5d4
+acfe7ee96efdec46b361f69680cf12b6dfbfb0c55c550161884a11bfa0ec56f0
+82b7b73d76367065dc9b4d4db5f1f33d595927d79c0e4066cffe18b6ac2a98a7
+01b9fe8cd285c172b6d84081bb30915a69a9d8916df7797994a9011df0bfba5c
+0f175f3a6e5f1ff2ea56ae48647026a99eccb27e65d77e468d7f8b54ac07b548
+2dfc5dd7227ffe07161896538abdcb1a1e9829361139dd46fd618df57dc9e4e8
+1726c94dda6294ac4e2d352230a8da0a46ab511588df5c9daace098b38546b8b
+49a056a802e7a4d95bb5f111e5532d1359bfdfbc6dcfa4da8bb1ac7e97ddb232
+bc9baf909972d2523deed1a9eeab71dd12be350d189590a25d1462d4feb1f529
+c40e28bce89b939787ff987ba0df312eae732bc54a84840606b6d5ada8f7cda2
+f6dbda3cfa7d8568ee70fb695ee3317dd2a603d7fadc00a8c0a4f4a9c7cb14eb
+2e43b8302aa869d64414f2bc36f60217aafec172d5ed2e37426ba73e7c5a083b
+8829dd23d2b12bb1b93da0314e161136eab64fa0d39054857d875cf5a3e31818
+03c9a65f3e62b348b227ac2cbc18d3f66af85ab5fd7a64148dd59d84967d3ca3
+7b847e05a12429d8a026ae8cfa9bb2bf7f9e2371fcec5ad897a9e2fcf082f511
+699a1df36a11d6df8b1131c8cd23fb14dc685f265abe77f66cc1b29f18367a20
+247361c00274849464f6c19b068f154ea3a81f4db4f2aed0ec93ce94841db68d
+081c5ce79da38d1ab42955e489996000b986bf51fbdf842151199c049b3debe9
+865a44c6a0e98508eb3a136f66d41fdc35ffaa174ab122406dd234733f0ba2d9
+83b8869165b2e11092202af599a9c6ad17946bb956e7ad7e0e7abe758ca14a23
+001ef0bc0cf08de5c8052c90b3ee7b5e37c53bdb34913eb92d6a198c973ac1f4
+b63a3a0cd1356fd37ec946162485a9bc85e91b8ad873945c23da14ce7369ffb0
+798055e28152eb8ac9a6648e04e934613f1c070e5100d1f2ad64b1019d08c55b
+faa74a6ecea415f4fc2ef2508acbc561d382748b24a5dee09d2a8e8ee1b8ce23
+a4563186008acc6ffa82216ea49c070a61a4c4770580256e4bff850b5d215b8b
+1da62c8f2820516d097d383107015cc4f1d7fd73b84dc12874403526d6d07270
+73e178c7fedb4306501f08146cdf100b59ff1440c6a50f1dab85b8439262618d
+d02750fc9c563a9a625868ac4ac7cdb3472915b71e12a674259b5216829e82e9
+f881ec03911cf3d09eb1460354931abf5e901a44d83500082493c643eb1c56b1
+1a6c8cac6ba750e3b66b536f53be4dbe5f8727108de82db7a5a6a8f24a5e318e
+02516a7d5d1519e8cacb0c27b6ade269653a0851be900dbcd39cefa05ea29d86
+e2f1d32e802f78d81bf654fd696f52704367641abb98e94a8e2bba4cc388fb5f
+f5666cc8f1dde43f198f37f92581e101252e8c7a30c9930869a9d0b8347e5f98
+3b0c147294a3ef6a55cacd922ed9bc844f453ad80c781571521d12a8eb856dfe
+ed28aed496e994c0fd3c8468ae403bdef36a74a76501a6315ba7f9d46e5d4b7f
+3bf48e711dbe61e7e3e3c1ad862fdaad994ecd3179336fb77ca29821a4a2d87d
+2434fa7370248d0381fcdfa1e47e275cede633a702c6c1e82712a279cebc5f38
+9b5519f670a7818c8ee482966e3659a5d3557a87c655eaa6db267c50a3d1d252
+67554995c31a50bd4452b385cba51ba6ae66f1ff5290caa7b4fd28364f08be27
+d8d515a0fc61cc4325635fdc6752f5a39571e2244f0d9f0ce15e8d47f608eafa
